+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+#### 0.5.3
+
+-   Added product fetching by SKU.
+
+#### 0.5.2
+
+-   Improved typing for `ListResult` model.
+-   Added searching for categories.
+
+#### 0.5.1
+
+-   Added `listByCategory(categoryId)` to products client.
+
+### 0.5.0
+
+-   Added `getProducts(categoryId)` to categories client.
+
 #### 0.4.2
 
 -   Fixed bug in `CategoryListResult.flattenCategoryListResults()` which caused it to return empty array always.